requirements.txt
dbus-python>=1.2.16
requests>=2.26.0
